Extract AppLogs endpoint tests to a separate file
